Fix php::config by making sure that $phpinidir exists
This required creating a new class php::common that php::solaris and
friends inherit from, and setting up a package alias for php-cli since
this is the closest thing to a base package name that PHP has across
platforms.

Additionally, the default provider for Solaris is set to pkgutil since
the Solaris code all requires OpenCSW packages.